% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{maximize_clock_tree_log_likelihood}
\alias{maximize_clock_tree_log_likelihood}
\title{Tree reconstruction via maximum parsimony}
\usage{
maximize_clock_tree_log_likelihood(
  states_alignment,
  tree,
  Q,
  pi,
  initial_depth
)
}
\arguments{
\item{states_table}{Matrix or data frame containing numericals representing an alignment of states with sites as rows and species/strains as columns}

\item{rooted_trees}{List of rooted tree topologies to be analyzed}
}
\description{
Determines the best fitting tree topology by maximum parsimony for given data
}
\examples{
test
}
