% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{maximum_parsimony}
\alias{maximum_parsimony}
\alias{all_parsimony_rooted_trees}
\alias{small_parsimony}
\title{Tree reconstruction via maximum parsimony}
\usage{
maximum_parsimony(states_table, rooted_trees)

all_parsimony_rooted_trees(states_table, rooted_trees)

small_parsimony(states_table, rooted_tree)
}
\arguments{
\item{states_table}{Matrix or data frame containing numericals representing an alignment of states with sites as rows and species/strains as columns.}

\item{rooted_trees}{List of rooted tree topologies of class phylo (ape) to be analyzed.}

\item{rooted_tree}{Rooted tree topology of class phylo (ape) to be analyzed.}
}
\description{
\code{maximum_parsimony} determines the single, best fitting among the given tree topologies by maximum parsimony for given data. The respective determined costs wil be attached to the phylo object.\cr\cr
\code{all_parsimony_rooted_trees} performs parsimony for all tree topologies given. The respective determined costs wil be attached to the phylo objects.\cr\cr
\verb{small_parsimony performs} parsimony for one tree topology. The respective determined costs wil be attached to the phylo object.
}
\examples{
test
}
