% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{simulate_evolution}
\alias{simulate_evolution}
\title{Generation of a states alignment}
\usage{
simulate_evolution(
  nstates,
  tree,
  Q,
  pi,
  noise_sd = 0,
  discretization = NA,
  small_num = 1e-04
)
}
\arguments{
\item{nstates}{Number of sites to be simulated (= number of rows of resulting states table).}

\item{tree}{Phylogenetic tree of class phylo used for simulation of evolution (branch lengths required!).}

\item{Q}{Numerical n x n matrix representing the substittion rate matrix of the evolutionary model to be used, with n being the number of states in the model.}

\item{pi}{Numerical vector of size n representing the equilibrium frequency, with n being the number of states in the model.}
}
\description{
This function simulates evolution given an evolutionary model and a phylogenetic tree resulting in the generation of a states table (alignment). It is possible to add a noise to the final result that could represent, e.g., sequencing errors.
}
\examples{
test
}
